pyproject.toml
[tool.poetry]
name = "iart_indexer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9.*"
redisai = "^1.0.1"
ml2rt = "^0.2.0"
imageio = "^2.9.0"
elasticsearch = "^7.9.1"
grpcio = "^1.32.0"
grpcio-tools = "^1.32.0"
umap-learn = "0.5.1"
scikit-learn = "^1.3.0"
h5py = "^3.1.0"
msgpack = "^1.0.2"
packaging = "^20.8"
spacy = "^3.0.1"
ftfy = "^5.9"
regex = "^2021.3.17"
elasticsearch-dsl = "^7.3.0"
tqdm = "^4.62.3"
transformers = "^4.24.0"
qdrant-client = "1.1.1"
ray = {extras = ["serve"], version = "^2.8.1"}
requests = "^2.31.0"
torch = "^2.1.2"
open-clip-torch = "^2.23.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
isort = "^5.5.2"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
